LPS-121274 Move SidebarPanelIdContext to separated file and folder
