+++ modules/apps/site-navigation/site-navigation-admin-web/src/main/resources/META-INF/resources/site_navigation_menu_editor/components/AppLayout.js
@@ -14,31 +14,29 @@
 
 import classNames from 'classnames';
 import PropTypes from 'prop-types';
-import React, {useContext, useEffect, useMemo, useState} from 'react';
+import React, {useEffect, useMemo} from 'react';
+
+import {
+	useSetSidebarPanelId,
+	useSidebarPanelId,
+} from '../contexts/SidebarPanelIdContext';
 
 const DEFAULT_SIDEBAR_PANELS = [];
 
-const SetSidebarPanelIdContext = React.createContext(() => {});
-
-export const useSetSidebarPanelId = () => {
-	return useContext(SetSidebarPanelIdContext);
-};
-
 export const AppLayout = ({
 	contentChildren,
-	initialSidebarPanelId = null,
 	sidebarPanels = DEFAULT_SIDEBAR_PANELS,
 	toolbarChildren,
-	...props
 }) => {
-	const [sidebarPanelId, setSidebarPanelId] = useState(initialSidebarPanelId);
+	const setSidebarPanelId = useSetSidebarPanelId();
+	const sidebarPanelId = useSidebarPanelId();
 
 	const SidebarPanel = useMemo(
 		() =>
 			sidebarPanels?.find(
 				(sidebarPanel) => sidebarPanel.sidebarPanelId === sidebarPanelId
 			)?.component,
-		[sidebarPanels, sidebarPanelId]
+		[sidebarPanelId, sidebarPanels]
 	);
 
 	useEffect(() => {
@@ -47,11 +45,11 @@ export const AppLayout = ({
 		return () => {
 			handler.removeListener();
 		};
-	}, []);
+	}, [setSidebarPanelId]);
 
 	useEffect(() => {
 		setSidebarPanelId(null);
-	}, [sidebarPanels]);
+	}, [setSidebarPanelId, sidebarPanels]);
 
 	useEffect(() => {
 		if (SidebarPanel) {
@@ -60,7 +58,7 @@ export const AppLayout = ({
 	}, [SidebarPanel]);
 
 	return (
-		<SetSidebarPanelIdContext.Provider value={setSidebarPanelId}>
+		<>
 			<div className="bg-white component-tbar tbar">
 				<div className="container-fluid container-fluid-max-xl">
 					<div className="px-1 tbar-nav">{toolbarChildren}</div>
@@ -85,10 +83,10 @@ export const AppLayout = ({
 						}
 					)}
 				>
-					{SidebarPanel && <SidebarPanel {...props} />}
+					{SidebarPanel && <SidebarPanel />}
 				</div>
 			</div>
-		</SetSidebarPanelIdContext.Provider>
+		</>
 	);
 };
