+++ modules/apps/site-navigation/site-navigation-admin-web/src/main/resources/META-INF/resources/site_navigation_menu_editor/components/SidebarPanelContent.js
@@ -22,7 +22,7 @@ import {globalEval} from 'metal-dom';
 import PropTypes from 'prop-types';
 import React, {useEffect, useRef, useState} from 'react';
 
-import {useSetSidebarPanelId} from './AppLayout';
+import {useSetSidebarPanelId} from '../contexts/SidebarPanelIdContext';
 
 export const SidebarPanelContent = ({
 	contentRequestBody,
